Aligned Autofac factory
