--- src/ServiceProviderFactorySample/AutofacServiceProviderFactory.cs
@@ -1,36 +1,0 @@
-﻿using System;
-using Autofac;
-using Autofac.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace ServiceProviderFactorySample
-{
-    public static class AutofacExtensions
-    {
-        public static IServiceCollection AddAutofac(this IServiceCollection services)
-        {
-            return services.AddSingleton<IServiceProviderFactory<ContainerBuilder>, AutofacServiceProviderFactory>();
-        }
-
-        public static IWebHostBuilder UseAutofac(this IWebHostBuilder host)
-        {
-            return host.ConfigureServices(services => services.AddAutofac());
-        }
-    }
-
-    internal class AutofacServiceProviderFactory : IServiceProviderFactory<ContainerBuilder>
-    {
-        public ContainerBuilder CreateBuilder(IServiceCollection services)
-        {
-            var containerBuilder = new ContainerBuilder();
-            containerBuilder.Populate(services);
-            return containerBuilder;
-        }
-
-        public IServiceProvider CreateServiceProvider(ContainerBuilder containerBuilder)
-        {
-            return new AutofacServiceProvider(containerBuilder.Build());
-        }
-    }
-}
